Order imports by relevance
