+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@heroicons/react": "^2.1.5",
     "next": "14.2.15",
     "react": "^18",
     "react-dom": "^18",
