+++ src/app/(quiz)/question4/page.tsx
@@ -1,8 +1,9 @@
 "use client";
 
+import Link from "next/link"
+;
 import AnswerButton from "@/components/AnswerButton";
 import ProgressBar from "@/components/ProgressBar";
-import Link from "next/link";
 
 const answerArray = [
   { label: "a. Breakage", value: "goals_anti-ageing" },
