+++ src/types/Quiz.ts
@@ -1,0 +1,7 @@
+export type QuizState = {
+    hairType: string;
+    washFrequency: string[];
+    desiredBenefit: string;
+    hairConcern: string;
+    hairColor: string;
+  };
