+++ src/types/tailwind-scrollbar-hide.d.ts
@@ -1,1 +1,1 @@
-declare module 'tailwind-scrollbar-hide';
+declare module "tailwind-scrollbar-hide";
